% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{lmix}
\alias{lmix}
\alias{dmix}
\alias{rmix}
\title{General Finite Mixture Distribution}
\usage{
lmix(x, w, ldens, ..., shared_args = NULL)

dmix(x, w, ldens, ..., shared_args = NULL, log = FALSE)

rmix(n, w, rdens, ..., shared_args = NULL)
}
\description{
General Finite Mixture Distribution
}
