% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_distributions.R
\name{lmix_mvtnorm}
\alias{lmix_mvtnorm}
\alias{dmix_mvtnorm}
\alias{rmix_mvtnorm}
\alias{ulmix_mvtnorm_temp}
\title{Finite Mixture of Multivariate Normal distributions}
\usage{
lmix_mvtnorm(x, w, ..., shared_args = NULL)

dmix_mvtnorm(x, w, ..., shared_args = NULL, log = FALSE)

rmix_mvtnorm(n, w, comp_args_list, shared_args = NULL)

ulmix_mvtnorm_temp(x, beta = 1, w, ..., shared_args = NULL)
}
\arguments{
\item{x}{A quantile vector or a matrix of quantile vectors (by rows).}

\item{w}{A vector of non-negative mixture weights. To be a valid mixture they must sum to 1.}

\item{...}{Other arguments passed on to \code{lmvtnorm} that vary by mixture component
(see Component Parameters)}

\item{shared_args}{List of other arguments passed on to \code{lmvtnorm} who are shared by all
mixture components (see Component Parameters).}

\item{log}{For \code{dmix_mvtnorm}, whether to return the log-density or not (the default).}

\item{n}{Number of observations.}

\item{comp_args_list}{For \code{rmix_mvtnorm}, a list whose k-th element contains the list of
parameters specific to the k-th mixture component (see Component Parameters).}

\item{beta}{Inverse temperature parameter \eqn{\beta > 0}.}
}
\value{
\code{lmix_mvtnorm} gives the log-density, \code{dmix_mvtnorm} the density and
\code{rmix_mvtnorm} generates random samples. Finally, \code{ulmix_mvtnorm_temp} gives the
unnormalized log-density of the tempered mixture.
}
\description{
Log-density, density, random sampling from a finite mixture of multivariate normal distributions,
as well as unnormalized log-density of the tempered mixture.
}
\details{
The tempering in \code{ulmix_mvtnorm_temp} is done \emph{after} the mixture; this is not to be confused
with a mixture of tempered multivariate normals. This latter scenario would just be a
regular mixture of multivariate normals, corresponding to \code{lmix_mvtnorm},
as tempering a multivariate normal just rescales it (see \link{lmvtnorm_temp}).
}
\section{Component Parameters}{
See the documentation for \link{lmix} for a lengthier exposition of the difference between passing
arguments via \code{...} or \code{shared_args}, while the documentation of \link{lmvtnorm} explains the specific
parameters available for the multivariate normal components of the mixture. These may typically
be lists for \code{mu} and \code{sigma}, passed through the \code{...} argument. But in case the scales are
the same for each component (or for the common dimension parameter \code{d}), the \code{shared_args}
list would be useful.
}

