% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{lmix_norm}
\alias{lmix_norm}
\alias{dmix_norm}
\alias{rmix_norm}
\alias{ulmix_norm_temp}
\title{Finite Mixture of Normal distributions}
\usage{
lmix_norm(x, w, mean, sd = NULL, shared_sd = 1)

dmix_norm(x, w, mean, sd = NULL, shared_sd = 1, log = FALSE)

rmix_norm(n, w, mean, sd = NULL, shared_sd = 1)

ulmix_norm_temp(x, beta = 1, w, mean, sd = NULL, shared_sd = 1)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{w}{A vector of non-negative mixture weights. To be a valid mixture they must sum to 1.}

\item{mean}{Vector containing the means of the components}

\item{sd}{(Optional) Vector containing the standard deviations of the components}

\item{shared_sd}{Shared standard deviation of the components; for different scales use \code{sd}.}

\item{log}{For \code{dmix_norm}, whether to return the log-density or not (the default).}

\item{n}{Number of observations.}

\item{beta}{Inverse temperature parameter \eqn{\beta > 0}.}
}
\value{
\code{lmix_norm} gives the log-density, \code{dmix_norm} the density and \code{rmix_norm} generates random samples.
Finally, \code{ulmix_norm_temp} gives the unnormalized log-density of the tempered mixture.
}
\description{
Log-density, density, random sampling from a finite mixture of normal distributions,
as well as unnormalized log-density of the tempered mixture.
}
\details{
The tempering in \code{ulmix_norm_temp} is done \emph{after} the mixture; this is not to be confused
with a mixture of tempered normals. This latter scenario would just be a regular mixture of normals,
corresponding to \code{lmix_norm}, as tempering a normal just rescales it (see \link{lnorm_temp}).
}
