% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lmvtnorm_temp}
\alias{lmvtnorm_temp}
\alias{dmvtnorm_temp}
\alias{rmvtnorm_temp}
\alias{lmvtnorm}
\alias{rmvtnorm}
\title{Multivariate Tempered Normal log densities and samplers}
\usage{
lmvtnorm_temp(
  x,
  beta = 1,
  mu = 0,
  sigma = NULL,
  sigma_inv = NULL,
  logdet_sigma = NULL,
  d = NULL
)

dmvtnorm_temp(
  x,
  beta = 1,
  mu = 0,
  sigma = NULL,
  sigma_inv = NULL,
  logdet_sigma = NULL,
  d = NULL
)

rmvtnorm_temp(n, beta = 1, mu, sigma = NULL, LChol_sigma = NULL, d = NULL)

lmvtnorm(x, mu = 0, sigma = NULL, sigma_inv = NULL, logdet_sigma = NULL)

rmvtnorm(n, mu = 0, sigma = NULL, LChol_sigma = NULL)
}
\description{
Multivariate Tempered Normal log densities and samplers
}
