% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_distributions.R
\name{lskewnorm}
\alias{lskewnorm}
\alias{lhatskewnorm}
\alias{lmix_hatskewnorm}
\alias{ulmix_hatskewnorm_temp}
\title{Skew Normal log-density and mixtures}
\usage{
lskewnorm(x, m, s, a)

lhatskewnorm(x, mu, sigma, alpha)

lmix_hatskewnorm(x, w = rep(0.25, 4), ..., shared_args = list(alpha = 2))

ulmix_hatskewnorm_temp(
  x,
  beta = 1,
  w = rep(0.25, 4),
  ...,
  shared_args = list(alpha = 2)
)
}
\arguments{
\item{x}{Vector of quantiles for \code{lskewnorm} or single multivariate quantile for \code{lhatskewnorm}}

\item{m}{Location parameter of univariate skew normal}

\item{s}{Positive scale parameter of univariate skew normal}

\item{a}{Skewness parameter of univariate skew normal}

\item{mu}{Location vector parameter of multivariate independent skew normal}

\item{sigma}{Positive scale vector parameter of multivariate independent skew normal}

\item{alpha}{Skewness vector parameter of multivariate independent skew normal}

\item{w}{A vector of non-negative mixture weights. To be a valid mixture they must sum to 1.}

\item{...}{Other arguments passed on to \code{lhatskewnorm} that vary by mixture component
(see the Component Parameters).}

\item{shared_args}{List of other arguments passed on to \code{lhatskewnorm} who are shared
by all mixture components (see Component Parameters).}

\item{beta}{Inverse temperature parameter \eqn{\beta > 0}.}
}
\value{
\code{lskewnorm} returns the log-density of a univariate skew normal, \code{lhatskewnorm} the log-density of
a product form multivariate skew normal (i.e. sum of independent skew normal log-densities), while
\code{lmix_hatskewnorm} and \code{ulmix_hatskewnorm_temp} return the normalized log-density and unnormalized
tempered log-density of a mixture of product form multivariate skew normals respectively.
}
\description{
Log-density of univariate skew normal distribution and multivariate version-
but with independent components- as well as mixture and unnormalized tempered mixture
of this multivariate product form of skew normals.
}
\details{
The skew normal distribution is parametrized by a real location, positive scale and real skewness
parameters denoted \eqn{m,\,s,\,a} respectively with the following density
\deqn{\tilde{\phi}_{sn}(x | m, s, a) = \dfrac{2}{s}\,\phi\left(\dfrac{x-m}{s}\right)\Phi\left[a\,\dfrac{x-m}{s}\right],}
where \eqn{\phi} and \eqn{\Phi} are the density and CDF of a standard normal random variable.

The product form of skew normals implemented here is then parametrized by corresponding vectors of
the above three parameters denoted \eqn{\mu, \sigma, \alpha} to offer maximum flexibility in the behaviour
of the components:
\deqn{f(x | \mu, \sigma, \alpha) = \prod\limits_{i=1}^d \tilde{\phi}_{sn}(x_i | \mu_i, \sigma_i, \alpha_i)}

The \eqn{K}-component mixture density and tempered mixture unnormalized density are accordingly
defined on top of the above multivariate product form:
\deqn{f_K(x | w, \theta) = \sum\limits_{k=1}^K w_k\,f(x | \mu_k, \sigma_k, \alpha_k) = \sum\limits_{k=1}^K w_k\,\prod\limits_{i=1}^d \tilde{\phi}_{sn}(x_i | \mu_{k,i}, \sigma_{k,i}, \alpha_{k,i})}
\deqn{\tilde{f}_K(x | w, \theta, \beta) \propto f_K(x | w, \theta)^\beta}
where to simplify notation we gather all the mixture component parameters \eqn{\mu, \sigma, \alpha}
into the single parameter \eqn{\theta}.
}
\section{Component Parameters}{
Same behavior as \link{lmix}: \code{...} expects iterable arguments to be swept across all components while
\code{shared_args} is a list of shared arguments across components.
For a more detailed explanation please see the documentation and examples in \link{lmix}.
}

\seealso{
\link{lmix} \link{ulmix_norm_temp}
}
