% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Metropolis_Hastings_Chains_Cpp.R
\name{rwm_global_scale_sampler_leaner_chain}
\alias{rwm_global_scale_sampler_leaner_chain}
\title{Random-Walk Metropolis}
\usage{
rwm_global_scale_sampler_leaner_chain(
  l_target,
  d,
  ...,
  global_scale = 2.38,
  S = 1000,
  burn = 0,
  x_0 = NULL,
  x_0_u = 2,
  l_0 = NULL,
  seed = NULL,
  custom_rw_sampler = NULL,
  more_sampler_args = NULL,
  silent = FALSE
)
}
\arguments{
\item{l_target}{Function which returns the log-density of the target distribution (possibly up to a constant)
and which accepts as first argument the state of the chain.}

\item{d}{Integer specifying the dimensionality of the state vector.}

\item{...}{Additional parameters passed to \code{l_target}.}

\item{global_scale}{Scale for the RWM sampler to use.}

\item{S}{Number of total samples to run the Markov Chain for.}

\item{burn}{Number of simulated samples to discard as burn-in. If \code{0} (the default),
the initial state \code{x_0} is discarded but otherwise all the simulated samples are kept.
Use \code{-1} to also keep he initial state \code{x_0}.}

\item{x_0}{(Optional) Initial state of the chain.}

\item{x_0_u}{(Optional) Positive number indicating the upper limit of a Uniform distribution
used to sample an initial value for the chain. The lower limit is then its negative. Hence,
each component of \code{x_0} is sampled uniformly between -\code{x_0_u} and \code{x_0_u}.
Ignored if \code{x_0} is present.}

\item{l_0}{(Optional) Precomputed value of \code{l_target} evaluated at \code{x_0}.}

\item{seed}{(Optional) Seed to set before starting sampling the chain.}

\item{custom_rw_sampler}{(Optional) Function for a user defined Random-Walk sampler.
It must accept as its first argument the current state of the chain and as its second the \code{global_scale}
used for the chain. If \code{NULL} (the default) the chain assumes a Normal RW sampler, see \code{Details}.}

\item{more_sampler_args}{(Optional) List containing further arguments passed to \code{custom_rw_sampler}.}

\item{silent}{(Logical) Should message-printing be disabled? By default, it is \code{FALSE}.}
}
\value{
A list containing the results of the RWM Chain:
list(x = x, l_x_curr = l_x, acc = acc, acc_rate = acc_rate)
\itemize{
\item \code{x}:
A matrix containing the evolution of the states of the chain with \code{d} rows and \code{S} - \code{burn}
columns.
\item \code{l_x_curr}:
Current (last) value of the log-density of the target.
\item \code{acc}:
A vector specifying whether or not each proposal was accepted or rejected (excluding burn-in).
\item \code{acc_rate}:
Observed acceptance rate across all iterations (including burn-in).
}
}
\description{
A single global scale real-valued parameter is assumed, dimension must be explicitly passed.
Less checks as this is mainly for developing ALPS internally
}
