% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings_Chains.R
\name{rwm_sampler_chain}
\alias{rwm_sampler_chain}
\title{Random Walk Metropolis Chain}
\usage{
rwm_sampler_chain(
  l_target,
  ...,
  scale = 1,
  S = 1000,
  burn = 0,
  x_0 = NULL,
  x_0_u = 2,
  l_0 = NULL,
  seed = NULL,
  custom_rw_sampler = NULL,
  more_sampler_args = NULL,
  d = NULL,
  silent = FALSE
)
}
\arguments{
\item{l_target}{Function which returns the log-density of the target distribution (possibly up to a constant)
and which accepts as first argument the state of the chain.}

\item{...}{(Optional) further arguments passed to \code{l_target}.}

\item{scale}{Scale of the RWM proposal (see \verb{RW Sampler})}

\item{S}{Number of samples of the chain, including burn-in but excluding starting state.}

\item{burn}{(Default = 0) Number of burn-in iterations. By default, the starting state of the chain
is always burned; if you would like to keep it, set this argument to -1 (see \code{Details}).}

\item{x_0}{(Optional) Starting state of the chain.}

\item{x_0_u}{(Default = 2) If \code{x_0} is not specified, then the starting state of the chain is randomly sampled
from a uniform distribution on (-\code{x_0_u},\code{x_0_u}). If \code{x_0} is present, it is ignored.}

\item{l_0}{(Optional) Precomputed value of the log-density at \code{x_0}.}

\item{seed}{(Optional) Seed for reproducibility of the algorithm passed on to \code{set.seed()}.}

\item{custom_rw_sampler}{(Optional) User-specified Random Walk sampler function, must accept as its first two arguments
the current state of the chain and \code{scale} (see \verb{RW Sampler}).}

\item{more_sampler_args}{(Optional) A list of further arguments passed on to \code{custom_rw_sampler}.}

\item{d}{(Optional) Dimensionality of the problem (see \code{Details}).}

\item{silent}{(Default = FALSE) Should the algorithm avoid printing messages?}
}
\value{
A list containing the results of the Random-Walk Metropolis Chain:
\itemize{
\item \code{x}:
A matrix whose rows contain the samples of the chain (after burn-in).
\item \code{l_x}:
A vector of log-densities of \code{x}.
\item \code{acc}:
A vector specifying whether or not each of the proposals was accepted or rejected.
\item \code{acc_rate}:
Acceptance rate of the chain.
\item \code{y}:
A matrix whose rows contain the proposals made during the chain (excluding burn-in).
\item \code{l_y}:
A vector of log-densities of \code{y}.
\item \code{delta_l}:
A vector of MH log-ratios.
}
}
\description{
Function for running a single chain using the Random Walk Metropolis algorithm for a user-specified target.
}
\details{
Missing details for dimension and burn-in
}
\section{RW Sampler}{
Missing section explaining behavior of \code{scale} and default/custom samplers.
}

