% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulated_Tempering.R
\name{ST_rwm_chain}
\alias{ST_rwm_chain}
\title{Simulated Tempering}
\usage{
ST_rwm_chain(
  l_target,
  ...,
  beta_schedule,
  g_schedule = NULL,
  scale = 1,
  Temp_Moves = 1000,
  Within_Moves = 10,
  burn_cycles = 0,
  x_0 = NULL,
  x_0_u = 2,
  k_0 = NULL,
  l_0 = NULL,
  seed = NULL,
  custom_rw_sampler = NULL,
  target_names = NULL,
  d = NULL,
  silent = FALSE
)
}
\arguments{
\item{l_target}{A function returning the log-density of the target (possibly up to a normalizing constant)
which is suitable for tempering. This is, its first argument should accept the state and its second argument
should accept an inverse-temperature parameter beta.}

\item{...}{(Optional) Further arguments passed on to \code{l_target}}

\item{beta_schedule}{The vector of inverse-temperatures to be used for tempering.}

\item{g_schedule}{(Optional) A vector of Simulated Tempering auxiliary constants.
If ommited, they are all fixed to 0.}

\item{scale}{Scale parameter for the Random Walk Metropolis within-temperature exploration moves.}

\item{Temp_Moves}{Number of temperature moves to attempt}

\item{Within_Moves}{Number of within-temperature exploration moves to conduct after each temperature move.}

\item{burn_cycles}{Burn-in cycles}

\item{x_0}{(Optional) Starting state for the chain}

\item{x_0_u}{(Default = 2) Whenever \code{x_0} is not specified, a starting state is
uniformly sampled between -\code{x_0_u} and \code{x_0_u} (independently in each dimension).}

\item{k_0}{(Optional) Starting temperature index.}

\item{l_0}{(Optional) Precomputed log-density value at the starting state and temperature level.}

\item{seed}{(Optional) Seed to be set using \code{set.seed()}.}

\item{custom_rw_sampler}{(Optional) User-specified RWM sampler(s) for the within-temperature moves.
See \code{Details} for more.}

\item{target_names}{(Optional) Names for the states components.}

\item{d}{(Optional) Dimensionality of the problem.}

\item{silent}{(Default = FALSE) Should the algorithm avoid printing messages?}
}
\description{
Simulated Tempering
}
