% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempeRing-package.R
\docType{package}
\name{tempeRing-package}
\alias{tempeRing}
\alias{tempeRing-package}
\title{tempeRing: Tempering MCMC Methods}
\description{
The goal of tempeRing is to provide an R package for my Summer Tempering project at the Warwick Statistics CDT. The package allows a user to run several MCMC methods, with a focus in Parallel Tempering and Annealing for multimodal targets. Available methods include both non-reversible and reversible swap dynamics for Parallel Tempering/Replica Exchange and Annealed Leap-Point Sampling (ALPS), as well as Simulated Tempering and Random Walk Metropolis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fazepher/tempeRing}
  \item Report bugs at \url{https://github.com/fazepher/tempeRing/issues}
}

}
\author{
\strong{Maintainer}: Fernando A. Zepeda Herrera \email{Fernando.Zepeda@warwick.ac.uk} (\href{https://orcid.org/0000-0001-6670-6303}{ORCID})

}
\keyword{internal}
