% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ulmix_hatskewnorm_temp}
\alias{ulmix_hatskewnorm_temp}
\title{Mixture of Skew Normals from HAT and ALPS papers}
\usage{
ulmix_hatskewnorm_temp(
  x,
  beta = 1,
  w = rep(0.25, 4),
  ...,
  shared_args = list(alpha = 2)
)
}
\description{
Mixture of Skew Normals from HAT and ALPS papers
}
