% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings.R
\name{mh_step}
\alias{mh_step}
\title{Metropolis-Hastings Step}
\usage{
mh_step(
  x_curr,
  x_prop,
  l_curr,
  l_prop,
  lq_c2p = 0,
  lq_p2c = 0,
  do_checks = TRUE
)
}
\arguments{
\item{x_curr, x_prop}{Two vectors of current and proposed states, whose length determines the
dimension of the state space. Alternatively, two matrices containing in its rows the states,
and in this case the number of columns determines the dimension.}

\item{l_curr, l_prop}{Vectors of values of log-densities, possibly up to proportionality, of the
current and proposed states.}

\item{lq_c2p, lq_p2c}{(Optional) Vectors of transition log-densities, possibly up to proportionality.
The suffix \verb{_c2p} corresponds to transitions from the \strong{c}urrent towards the \strong{p}roposed state(s),
while the suffix \verb{_p2c} contains the reversed transitions. If these values are ommited (the default),
the Hastings ratio becomes standard Metropolis.}

\item{do_checks}{If TRUE (the default), run preliminary checks for arguments validity.}
}
\value{
A list containing the results of the Metropolis-Hastings step:
\itemize{
\item accepted: Whether or not the proposal was accepted or rejected.
\item x_next: The next value of the chain. If \code{accepted}, the proposed state, otherwise the current.
\item l_next: The value of the log-density of \code{x_next}.
}
}
\description{
Metropolis-Hastings Step
}
