% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings.R
\name{mh_step}
\alias{mh_step}
\title{Metropolis-Hastings Step}
\usage{
mh_step(
  x_curr,
  x_prop,
  l_curr,
  l_prop,
  lq_c2p = 0,
  lq_p2c = 0,
  do_checks = TRUE
)
}
\arguments{
\item{x_curr}{A vector or matrix of current state(s)}

\item{x_prop}{A vector or matrix of proposed state(s)}

\item{l_curr}{A vector of current log-densities, possibly up to proportionality}

\item{l_prop}{A vector of proposed log-densities, possibly up to proportionality}

\item{lq_c2p}{A vector of transition log-densities from the current state(s) to the proposal(s)}

\item{lq_p2c}{A vector of transition log-densities from the porposal(s) to the currrent state(s)}

\item{do_checks}{If TRUE (the default), run preliminary checks for validity of arguments}
}
\value{
A list containing the result of the Metropolis-Hastings step
}
\description{
Metropolis-Hastings Step
}
