% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{ulmix_temp}
\alias{ulmix_temp}
\alias{lmix_temp}
\title{Tempered Mixture log densities generator}
\usage{
ulmix_temp(x, beta = 1, w, ldens, ..., shared_args = NULL)

lmix_temp(x, beta = 1, w, ldens, ..., shared_args = NULL, log_z = NULL)
}
\arguments{
\item{x}{A quantile or quantiles if \code{ldens} allows for vectorization.}

\item{beta}{Inverse temperature parameter β > 0.}

\item{w}{A vector of non-negative mixture weights. To be a valid mixture they must sum to 1.}

\item{ldens}{A function that returns the log-density of the desired common mixture distribution.
For \code{lmix_temp} see Details for restrictions on multivariate densities.}

\item{...}{Other arguments passed on to \code{ldens} that vary by mixture component
(see Component Parameters).}

\item{shared_args}{List of other arguments passed on to \code{ldens} who are shared
by all mixture components (see Component Parameters).}

\item{log_z}{For \code{lmix_temp}, normalizing constant. If NULL (the default) it is estimated
via \code{stats::integrate}. Hence, while optional for univariate mixtures, it is necessary
for proper behavior on multivariate mixtures (see Details).}
}
\value{
\code{ulmix_temp} returns the unnormalized log-density of the mixture, while \code{lmix_temp} estimates
the normalized value.
}
\description{
Unnormalized and normalized log-tempered densities from a mixture of a user defined
\strong{log-density}.

While mixing densities still yields a normalized density, tempering \emph{does not} do so
and requires renormalization. Hence, the direct result of tempering is the unnormalized
\code{ulmix_temp}. The normalized version \code{lmix_temp} is provided mainly for \strong{univariate} log-densities,
where internal numerical integration is used via \code{stats::integrate}.
\subsection{Component Parameters}{

Same behavior as \link{lmix}: \code{...} expects iterable arguments to be swept across all components while
\code{shared_args} is a list of shared arguments across components.
For a more detailed explanation please see the documentation and examples in \link{lmix}.
}
}
\details{
When using \code{lmix_temp} for general \emph{multivariate} mixtures, the user should provide the
normalizing constant \code{log_z} (in the log scale). This parameter can also be useful for efficiency in
one-dimensional situations, as one would avoid estimating it at every call.
Note, however, that many tasks in the context of the package don't necessarily require the
normalized version; this function is provided for completeness and as a tool
for those situations were the normalized versions are indeed sought.
}
\seealso{
\link{lmix} \link{ulmix_norm_temp}
}
